+++ core/auth/src/main/kotlin/com/stash/core/auth/discord/DiscordHeaders.kt
